+++ lib/CSkillHandler.cpp
@@ -244,6 +244,7 @@ std::shared_ptr<CSkill> CSkillHandler::loadFromJson(const std::string & scope, c
 		skill->tags.emplace_back("spellSchool");
 
 	LIBRARY->generaltexth->registerString(scope, skill->getNameTextID(), json["name"]);
+	skill->offerCooldown = static_cast<int>(json["offerCooldown"].Integer());
 
 	for(auto skillPair : json["gainChance"].Struct())
 	{